Made parsing for constants "better"?
Distinct declarations will be highlighted as types but other type
aliases will be highlighted as constants. This is the best way to
resolve this, because we cant know there `Some` in `Other :: pkg.Some`
is some type or some simple constant. This is where an LSP is needed to
provide semantic tokens.